backend/requirements.txt
flask
flask-cors
langchain
langchain-openai
langchain-community
langchain-qdrant
qdrant-client
openai
python-dotenv
pypdf
werkzeug